requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
httpx==0.27.2
pydantic==2.9.2
qdrant-client==1.10.1
meilisearch==0.37.1
scikit-learn==1.5.2
numpy==2.1.2
pandas==2.2.3
python-dotenv==1.0.1
prometheus-fastapi-instrumentator==7.0.0
pyyaml==6.0.2
redis==5.0.7
sentence-transformers>=2.7.0
